+++ test/utils.js
@@ -54,6 +54,34 @@ describe('Utils', () => {
         {
           key: 'a',
           join: 'and',
+          children: [
+            {
+              key: 'b',
+              join: 'and',
+              children: [
+                {
+                  key: 'c',
+                },
+                {
+                  key: 'd',
+                  _meta: {
+                    filter: 'test',
+                  },
+                },
+              ],
+            },
+          ],
+        }
+      )
+      expect(result).to.deep.equal('test')
+    })
+    it('should handle basic sibling, but with items instead of children', () => {
+      let result = getRelevantFilters(
+        DebugProvider.groupCombinator,
+        ['a', 'b', 'c'],
+        {
+          key: 'a',
+          join: 'and',
           items: [
             {
               key: 'b',
@@ -82,11 +110,11 @@ describe('Utils', () => {
         {
           key: 'a',
           join: 'and',
-          items: [
+          children: [
             {
               key: 'b',
               join: 'and',
-              items: [
+              children: [
                 {
                   key: 'c',
                 },
@@ -118,11 +146,11 @@ describe('Utils', () => {
         {
           key: 'a',
           join: 'and',
-          items: [
+          children: [
             {
               key: 'b',
               join: 'and',
-              items: [
+              children: [
                 {
                   key: 'c',
                 },
@@ -154,11 +182,11 @@ describe('Utils', () => {
         {
           key: 'a',
           join: 'and',
-          items: [
+          children: [
             {
               key: 'b',
               join: 'or',
-              items: [
+              children: [
                 {
                   key: 'c',
                 },
@@ -188,11 +216,11 @@ describe('Utils', () => {
         {
           key: 'a',
           join: 'and',
-          items: [
+          children: [
             {
               key: 'b',
               join: 'and',
-              items: [
+              children: [
                 {
                   key: 'c',
                 },
@@ -207,7 +235,7 @@ describe('Utils', () => {
             {
               key: 'blah',
               join: 'not',
-              items: [
+              children: [
                 {
                   key: 'asdf',
                   _meta: {
@@ -236,11 +264,11 @@ describe('Utils', () => {
         {
           key: 'root',
           join: 'and',
-          items: [
+          children: [
             {
               key: 'criteria',
               join: 'and',
-              items: [
+              children: [
                 {
                   key: 'cgnya6ja8ys10iwl8fr',
                   _meta: {
@@ -250,7 +278,7 @@ describe('Utils', () => {
                 {
                   key: 'criteria',
                   join: 'or',
-                  items: [
+                  children: [
                     {
                       key: '8ilrqpm1je3m8ed5z5mi',
                       _meta: {
@@ -274,7 +302,7 @@ describe('Utils', () => {
               type: 'group',
               key: 'analysis',
               join: 'and',
-              items: [
+              children: [
                 {
                   key: 'results',
                 },
@@ -299,11 +327,11 @@ describe('Utils', () => {
         {
           key: 'root',
           join: 'and',
-          items: [
+          children: [
             {
               key: 'criteria',
               join: 'and',
-              items: [
+              children: [
                 {
                   key: 'cgnya6ja8ys10iwl8fr',
                   _meta: {
@@ -313,7 +341,7 @@ describe('Utils', () => {
                 {
                   key: 'criteria',
                   join: 'or',
-                  items: [
+                  children: [
                     {
                       key: '8ilrqpm1je3m8ed5z5mi',
                       _meta: {
@@ -336,12 +364,12 @@ describe('Utils', () => {
             {
               key: 'analysisOR',
               join: 'or',
-              items: [
+              children: [
                 {
                   type: 'group',
                   key: 'analysis',
                   join: 'and',
-                  items: [
+                  children: [
                     {
                       key: 'results',
                     },
@@ -374,11 +402,11 @@ describe('Utils', () => {
         {
           key: 'root',
           join: 'or',
-          items: [
+          children: [
             {
               key: 'criteria',
               join: 'and',
-              items: [
+              children: [
                 {
                   key: 'cgnya6ja8ys10iwl8fr',
                   _meta: {
@@ -388,7 +416,7 @@ describe('Utils', () => {
                 {
                   key: 'criteria',
                   join: 'or',
-                  items: [
+                  children: [
                     {
                       key: '8ilrqpm1je3m8ed5z5mi',
                       _meta: {
@@ -411,12 +439,12 @@ describe('Utils', () => {
             {
               key: 'analysisOR',
               join: 'and',
-              items: [
+              children: [
                 {
                   type: 'group',
                   key: 'analysis',
                   join: 'and',
-                  items: [
+                  children: [
                     {
                       key: 'results',
                     },
@@ -442,11 +470,11 @@ describe('Utils', () => {
         {
           key: 'root',
           join: 'and',
-          items: [
+          children: [
             {
               key: 'criteria',
               join: 'and',
-              items: [
+              children: [
                 {
                   key: 'cgnya6ja8ys10iwl8fr',
                   _meta: {
@@ -456,7 +484,7 @@ describe('Utils', () => {
                 {
                   key: 'criteria2',
                   join: 'or',
-                  items: [
+                  children: [
                     {
                       key: 'dod',
                       _meta: {
@@ -480,7 +508,7 @@ describe('Utils', () => {
               type: 'group',
               key: 'analysis',
               join: 'and',
-              items: [
+              children: [
                 {
                   key: 'results',
                   _meta: {